Extract the onDrugEnd function into a separate file.
