+++ src/components/preview/ui/Preview.jsx
@@ -9,6 +9,7 @@ import Header from "../components/Header";
 import LeftSide from "../components/LeftSide";
 import RightSide from "../components/RightSide";
 import A4PageWrapper from "../components/A4PageWrapper";
+import {onDragEndHandler} from "../utils/onDrugEndHandler";
 
 const DragDropContext = dynamic(
   () =>
@@ -30,67 +31,11 @@ const Preview = () => {
     {name: "website", icon: <CgWebsite/>},
   ];
 
-  const onDragEnd = (result) => {
-    const {destination, source} = result;
-
-    if (!destination) return;
-
-    if (
-      destination.droppableId === source.droppableId &&
-      destination.index === source.index
-    )
-      return;
-
-    if (source.droppableId === "work-experience") {
-      const newWorkExperience = [...resumeData.workExperience];
-      const [removed] = newWorkExperience.splice(source.index, 1);
-      newWorkExperience.splice(destination.index, 0, removed);
-      setResumeData({...resumeData, workExperience: newWorkExperience});
-    }
-
-    if (source.droppableId.includes("WORK_EXPERIENCE_KEY_ACHIEVEMENT")) {
-      const newWorkExperience = [...resumeData.workExperience];
-      const workExperienceIndex = parseInt(source.droppableId.split("-")[1]);
-      const keyAchievements =
-        newWorkExperience[workExperienceIndex].keyAchievements.split("\n");
-      const [removed] = keyAchievements.splice(source.index, 1);
-      keyAchievements.splice(destination.index, 0, removed);
-      newWorkExperience[workExperienceIndex].keyAchievements =
-        keyAchievements.join("\n");
-      setResumeData({...resumeData, workExperience: newWorkExperience});
-    }
-
-    if (source.droppableId === "skills") {
-      const newSkills = [...resumeData.skills];
-      const [removed] = newSkills.splice(source.index, 1);
-      newSkills.splice(destination.index, 0, removed);
-      setResumeData({...resumeData, skills: newSkills});
-    }
-
-    if (source.droppableId.includes("projects")) {
-      const newProjects = [...resumeData.projects];
-      const [removed] = newProjects.splice(source.index, 1);
-      newProjects.splice(destination.index, 0, removed);
-      setResumeData({...resumeData, projects: newProjects});
-    }
-
-    if (source.droppableId.includes("PROJECTS_KEY_ACHIEVEMENT")) {
-      const newProjects = [...resumeData.projects];
-      const projectIndex = parseInt(source.droppableId.split("-")[1]);
-      const keyAchievements =
-        newProjects[projectIndex].keyAchievements.split("\n");
-      const [removed] = keyAchievements.splice(source.index, 1);
-      keyAchievements.splice(destination.index, 0, removed);
-      newProjects[projectIndex].keyAchievements = keyAchievements.join("\n");
-      setResumeData({...resumeData, projects: newProjects});
-    }
-  };
-
   return (
     <div className="md:max-w-[60%] sticky top-0 preview rm-padding-print p-6 md:overflow-y-scroll md:h-screen">
       <A4PageWrapper>
         <ModalHighlightMenu/>
-        <DragDropContext onDragEnd={onDragEnd}>
+        <DragDropContext onDragEnd={onDragEndHandler}>
           <Header resumeData={resumeData} icons={icons}/>
           <hr className="border-dashed my-2"/>
           <div className="grid grid-cols-3 gap-6">
